add database postgres

+++ api/v1/campaign/models/campaign_objective.py
@@ -28,8 +28,8 @@ class CampaignObjective(models.Model):
     name = models.CharField(max_length=500, blank=False, null=False)
     description = models.CharField(max_length=800, blank=False, null=False)
     is_active = models.BooleanField(default=False)
-    created_by = models.IntegerField(null=True, blank=True)
-    updated_by = models.IntegerField(null=True, blank=True)
+    created_by = models.BigIntegerField(null=True, blank=True)
+    updated_by = models.BigIntegerField(null=True, blank=True)
     created_date = models.DateField(null=True, blank=True, default=datetime.now())
     updated_date = models.DateField(null=True, blank=True, default=datetime.now())
